requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
drf-yasg